requirements.txt
django>=4.0
djangorestframework
django-cors-headers
mysqlclient
cryptography
djangorestframework-simplejwt
python-dotenv
drf-spectacular
